+++ numerics/src/FrictionContact/gfc3d_ipm.c
@@ -236,7 +236,7 @@ static double getNewtonStepLength(const double * const x, const double * const d
 static void primalResidualVector(const double * velocity, NumericsMatrix * H,
                                  const double * globalVelocity, const double * w, double * out)
 {
-  double nd = H->size0;
+  size_t nd = H->size0;
 
   /* The memory for the result vectors should be allocated using calloc
    * since H is a sparse matrix. In other case the behaviour will be undefined.*/
@@ -262,7 +262,7 @@ static void dualResidualVector(NumericsMatrix * M, const double * globalVelocity
                                NumericsMatrix * H, const double * reaction, const double * f,
                                double * out)
 {
-  double m = H->size1;
+  size_t m = H->size1;
 
   /* The memory for the result vectors should be allocated using calloc
    * since H is a sparse matrix. In other case the behaviour will be undefined.*/
